adapt matlab transformer for summit recordings

+++ source/transformers/matlab.py
@@ -1,10 +1,12 @@
 import os
 import shutil
 import subprocess
-import matlab
-
+import matlab.engine
+from common.utils.time import unix_to_timestamps
+from common.utils.ingest import storage_format_date
 import numpy as np
 import pandas as pd
+import re
 
 
 from source.transformers.base import BaseTransformer
@@ -19,22 +21,37 @@ class OpenMindTransformerMixin(BaseTransformer):
         Convert the raw JSON files from the Medtronic summit RC+S API to anonymized CSV files using OpenMind code
         https://github.com/openmind-consortium/Analysis-rcs-data
         """
-        # TODO: move all of Raph's code here
 
-        print('Aggregate')
+        self.info("Aggregate data session")
         self.aggregate_data_sessions()
-        print('Anonymize')
-        self.anonymize_batch()
-        print('To csv')
-        self.convert_json_to_csv()
+        
+        #print('Anonymize')
+        #self.anonymize_batch()
+        
+        self.info("Convert JSON to CSV")
+        failures = self.convert_json_to_csv()
 
-        return None
+        failures_ = []
+        for f in failures:
+            failures.append({
+                "filename": f,
+                "type": "folder",
+            })
+
+        # ready = {'to upload': tasks['to do'], 'failure': tasks['failure']}
+        ready = {'to upload': [], 'failure': failures}
+
+        return ready
 
     def aggregate_data_sessions(self):
         # Organizes the session folders into date folders
-        directory = './temp/combined_original'
+        directory = self.source_location["path"]  # './temp/combined_original'
         unix_regex = 'Session([0-9]*)'
-        new_directory = directory[0:-8] + r'by_date'
+
+        #new_directory = directory[0:-8] + r'by_date'
+        new_directory = self.middle_location["path"]
+
+        file_cnt = 0
         for session_folder in os.listdir(directory):
             match = re.search(unix_regex, session_folder)
             if match:
@@ -42,28 +59,35 @@ class OpenMindTransformerMixin(BaseTransformer):
                 date_str = storage_format_date(date)
                 date_dir = os.path.join(new_directory, date_str)
                 os.makedirs(date_dir, exist_ok=True)
-                shutil.copytree(os.path.join(directory, session_folder), os.path.join(date_dir, session_folder))
+                if not os.path.exists(os.path.join(date_dir, session_folder)):
+                    shutil.copytree(os.path.join(directory, session_folder), os.path.join(date_dir, session_folder))
+                    if self.LIMIT_FILES_COPY != -1:
+                        file_cnt += 1
+                        if file_cnt > self.LIMIT_FILES_COPY:
+                            break
         return None
 
     def start_matlab(self):
         # Start and return matlab engine to run matlab code in python
-        return matlab.engine.start_matlab('-nojvm')
+        return 
 
     def anonymize_batch(self):
         # Run matlab anonymize_batch.m code
         eng = self.start_matlab()
-        eng.addpath(r'./', nargout=0)  # Path to matlab functions
+        eng.addpath(r'scripts/openmind', nargout=0)  # Path to matlab functions
         my_path = r'./temp/combined_by_date'
         eng.anonymize_batch_callable(my_path)
         return None
 
     def convert_json_to_csv(self):
-        # Run matlab convert_json_batch.m
-        eng = self.start_matlab()
-        eng.addpath(r'./', nargout=0)  # Path to data_analysis/openmind_processing folder
-        my_path = r'./temp/combined_anonymized_json'
-        eng.convert_json_batch_callable(my_path)
-        return None
+
+        my_path = self.middle_location["path"]
+    
+        with matlab.engine.start_matlab('-nojvm') as eng:
+            eng.addpath(r'scripts/openmind', nargout=0)                     
+            failures = eng.convert_json_batch_callable(my_path)
+
+        return failures
 
     def upload_to_wasabi(self):
         # Upload folders to wasabi.
